requirements.txt
fastapi>=0.101.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
web3>=7.0.0
eth-account>=0.13.0
openai>=1.0.0
httpx>=0.25.0
